some relatively intrusive changes in an attempt to understand calling.
In my initial runs of tnseq_gi, I noticed that the fdr estimates were
being returned in sorted order.  When I looked more into the code, I saw
that there are two functions at the top (bFDR and FWER_Bayes) which, to
my eyes, do a reasonable sortarg().sortarg() in order to calculate the
p-adjustment, but in the actual main() this is eschewed and a single
sortarg() operation is performed.  In my playing around to see if this
is indeed the problem, I added a statsmodels invocation of
multipletests() to see if my guesses were crazypants or not.

In any event, in the end I chose to change the code in main() to use
the functions at the top.  I hope this is correct.